Fixed issues with enemy counter
Enemy counter was finding objects with 'enemy' tag but no enemyHealth script, so the counter was going up to really high but half the entries were invalid.
Added line which only adds enemies to list if they have bothy the tag and the enemyhealth script.
Worth going through and seeing what objects are using the Enemy tag without enemyHealth to see if it's a mistake/why it has that tag.