[SUP 105] Quick fix for player health
Something in the game is sending the player a TakeDaamge message with a vlaue of 20, instantly killing the player. I can't find where this is coming from, so as a quick fix any damage values greater than 5 are set to 1, making the assumption they're using the old percentage based health system where a player has 100 health, not 10.